pyproject.toml
[tool.poetry]
name = "panops-data-registry"
version = "0.1.0"
description = "A registry of Earth observation datasets used in the PANOPS project of the Geosense lab at the University of Freiburg."
authors = ["Daniel Lusk"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "~3.12"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
ipykernel = "^6.29.0"
pygbif = "^0.6.4"
google-auth = "^2.29.0"
earthengine-api = "^0.1.395"
xee = "^0.0.11"
dask = {extras = ["distributed"], version = "^2024.3.1"}
pyyaml = "^6.0.1"
zarr = "^2.17.1"
tqdm = "^4.66.2"
rioxarray = "^0.15.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
pytest = "^7.4.4"
pylint = "^3.0.3"
pylint-pytest = "^1.1.7"
pytest-mock = "^3.12.0"
ruff = "^0.12.4"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests requiring external services",
]
filterwarnings = ["ignore::DeprecationWarning"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
